requirements.txt
streamlit==1.40.2
langchain==0.0.309
langchain-community==0.0.4
chromadb==0.3.26
ollama
sentence-transformers==2.2.2
pypdf==3.12.0
PyYAML==6.0.1
deep-translator==1.8.3
python-docx==0.8.11
beautifulsoup4==4.12.2
requests==2.31.0
langchain-ollama